+++ src/controllers/CompositeController.php
@@ -18,6 +18,7 @@ use blackcube\admin\actions\composite\CreateAction;
 use blackcube\admin\actions\composite\DeleteAction;
 use blackcube\admin\actions\composite\EditAction;
 use blackcube\admin\actions\composite\IndexAction;
+use blackcube\admin\actions\ExportAction;
 use blackcube\admin\actions\SeoAction;
 use blackcube\admin\actions\SitemapAction;
 use blackcube\admin\actions\TagAction;
@@ -85,7 +86,14 @@ class CompositeController extends BaseElementController
                         'file-preview', 'file-upload', 'file-delete',
                     ],
                     'roles' => ['@'],
-                ]
+                ],
+                [
+                    'allow' => true,
+                    'actions' => [
+                        'export',
+                    ],
+                    'roles' => [Rbac::PERMISSION_COMPOSITE_EXPORT],
+                ],
             ]
         ];
         $behaviors['forceAjax'] = [
@@ -140,6 +148,10 @@ class CompositeController extends BaseElementController
         ];
         $actions['delete'] = [
             'class' => DeleteAction::class,
+        ];
+        $actions['export'] = [
+            'class' => ExportAction::class,
+            'elementClass' => Composite::class,
         ];
         return $actions;
     }